Extract SharedAudioContext and SharedAudioTagsContext to distinct files
Splits `packages/core/src/audio/shared-audio-tags.tsx` into smaller, focused modules:
- `shared-audio-context.tsx`
- `shared-audio-tags-context.tsx`
- `shared-audio-tags-helpers.ts`

Updates imports across the repository.
